Add kernel and userspace callstack event context
LTTng callstack context. The callstack context can be added to any kernel
event. It records either the kernel or the userspace callstack, up to a
max depth. The context is a CTF sequence, such that it uses only the space
required for the number of callstack entries.

It allocates callstack buffers per-CPU and for each possible interrupt
context (4 interrupt nesting: none, softirq, irq, nmi). It therefore uses a
fixed amount of memory, proportional to the number of CPUs:

  size = cpus * nest * depth * sizeof(unsigned long)

Which is about 800 bytes per-CPUs on 64-bit host and a depth of 25. The
allocation is done at the initialization to avoid memory allocation
overhead while tracing.

The kernel callstack is recovered using save_stack_trace(), and the
userspace callstack uses save_stack_trace_user(). They rely on frame
pointers. These are usually available for the kernel, but the compiler
option -fomit-frame-pointer frequently used in popular Linux distributions
may cause the userspace callstack to be unreliable, and is a known
limitation of this approach. We still provide the feature, because it works
well for runtime environments having frame pointers.

The symbol name resolution is left to the trace reader.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>