Add kernel and userspace callstack event context
The 'csk' and 'csu' event contexts record the callstack of the kernel and
userspace program respectively when a kernel tracepoint is hit.

The kernel callstack shows the code location that triggered the event. The
callstack of userspace is recorded on the kernel event. For instance, if the
context is added to the sched_switch event, it records the callstack when the
program was removed from the CPU, either blocking or preempted.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>